Relocated Command creation code into the FrameProcessorPluginController class.
