+++ src/fastcs_odin/frame_processor.py
@@ -2,6 +2,7 @@ import re
 from collections.abc import Sequence
 
 from fastcs.attributes import AttrR
+from fastcs.cs_methods import Command
 from fastcs.datatypes import Bool, Int
 
 from fastcs_odin.odin_adapter_controller import (
@@ -78,6 +79,7 @@ class FrameProcessorPluginController(OdinAdapterController):
     """SubController for a plugin in a frameProcessor application."""
 
     async def initialise(self):
+        await self._create_commands()
         if any("dataset" in p.path for p in self.parameters):
 
             def __dataset_parameter(param: OdinParameter):
@@ -94,6 +96,25 @@ class FrameProcessorPluginController(OdinAdapterController):
                 await dataset_controller.initialise()
 
         return await super().initialise()
+
+    async def _create_commands(self):
+        command_path = f"command/{self.path[-1].lower()}"
+        command_response = await self.connection.get(
+            f"{self._api_prefix}/{command_path}/allowed"
+        )
+        if "allowed" in command_response:
+            command_names = command_response["allowed"]
+            assert isinstance(command_names, list)
+            for command_name in command_names:
+                self.construct_command(command_name, command_path)
+
+    def construct_command(self, command_name, command_path):
+        async def submit_command() -> None:
+            await self.connection.put(
+                f"{self._api_prefix}/{command_path}/execute", command_name
+            )
+
+        setattr(self, command_name, Command(submit_command))
 
     def _process_parameters(self):
         for parameter in self.parameters:
